[fix] typoしてるページをリネーム
