+++ view/next-project/src/pages/sponsors/index.tsx
@@ -1,18 +1,16 @@
-import { ChakraProvider } from '@chakra-ui/react';
-import { Table, Thead, Tbody, Tr, Th, Td, Flex, Spacer, Select } from '@chakra-ui/react';
-import { Center, Box } from '@chakra-ui/react';
+import clsx from 'clsx';
 import Head from 'next/head';
 import { RiAddCircleLine } from 'react-icons/ri';
 
-import theme from '@assets/theme';
+import { Card, Title } from '@components/common';
 import EditButton from '@components/common/EditButton';
-import Header from '@components/common/Header';
 import RegistButton from '@components/common/RegistButton';
+import MainLayout from '@components/layout/MainLayout';
 
 import type { NextPage } from 'next';
 
 const sponsorship: NextPage = () => {
-  const sponsorshipList = [
+  const sponsershipList = [
     {
       id: 1,
       name: '海龍',
@@ -20,129 +18,128 @@ const sponsorship: NextPage = () => {
       email: 'kairyu@gmail.com',
       address: '新潟県長岡市上富岡町2丁目280-1',
       representative: '長岡太郎',
-      created_at: '2022/3/1',
-      updated_at: '2022/3/2',
     },
   ];
   return (
-    <ChakraProvider theme={theme}>
+    <MainLayout>
       <Head>
-        <title>Create Next App</title>
-        <meta name='description' content='Generated by create next app' />
-        <link rel='icon' href='/favicon.ico' />
+        <title>協賛企業一覧</title>
+        <meta name='viewport' content='initial-scale=1.0, width=device-width' />
       </Head>
-
-      <Header />
-      <hr />
-
-      <Center>
-        <Box m='10' px='10' boxShadow='base' rounded='lg'>
-          <Box mt='10' mx='5'>
-            <Flex>
-              <Center mr='5' fontSize='2xl' fontWeight='100' color='black.0'>
-                企業一覧
-              </Center>
-              <Select variant='flushed' w='100'>
-                <option value='2021'>2021</option>
-                <option value='2022'>2022</option>
-              </Select>
-            </Flex>
-            <Flex>
-              <Spacer />
-              <Box>
-                <RegistButton>
-                  <RiAddCircleLine
-                    size={20}
-                    style={{
-                      marginRight: 5,
-                    }}
-                  />
-                  企業登録
-                </RegistButton>
-              </Box>
-            </Flex>
-          </Box>
-          <Box p='5' mb='2'>
-            <Table>
-              <Thead>
-                <Tr>
-                  <Th borderBottomColor='#76E4F7'>
-                    <Center fontSize='sm' color='black.600'>
-                      ID
-                    </Center>
-                  </Th>
-                  <Th borderBottomColor='#76E4F7'>
-                    <Center fontSize='sm' mr='1' color='black.600'>
-                      企業名
-                    </Center>
-                  </Th>
-                  <Th borderBottomColor='#76E4F7'>
-                    <Center fontSize='sm' color='black.600'>
-                      電話番号
-                    </Center>
-                  </Th>
-                  <Th borderBottomColor='#76E4F7' isNumeric>
-                    <Center fontSize='sm' color='black.600'>
-                      メール
-                    </Center>
-                  </Th>
-                  <Th borderBottomColor='#76E4F7'>
-                    <Center>住所</Center>
-                  </Th>
-                  <Th borderBottomColor='#76E4F7'>
-                    <Center color='black.600'>代表者</Center>
-                  </Th>
-                  <Th borderBottomColor='#76E4F7'>
-                    <Center color='black.600'></Center>
-                  </Th>
-                  <Th borderBottomColor='#76E4F7'>
-                    <Center color='black.600'>作成日時</Center>
-                  </Th>
-                  <Th borderBottomColor='#76E4F7'>
-                    <Center color='black.600'>更新日時</Center>
-                  </Th>
-                </Tr>
-              </Thead>
-              <Tbody>
-                {sponsorshipList.map((sponsorshipItem) => (
-                  <Tr key={sponsorshipItem.id}>
-                    <Td>
-                      <Center color='black.300'>{sponsorshipItem.id}</Center>
-                    </Td>
-                    <Td>
-                      <Center color='black.300'>{sponsorshipItem.name}</Center>
-                    </Td>
-                    <Td>
-                      <Center color='black.300'>{sponsorshipItem.tel}</Center>
-                    </Td>
-                    <Td isNumeric color='black.300'>
+      <Card w='w-1/1'>
+        <div className={clsx('mx-5 mt-10')}>
+          <div className={clsx('flex')}>
+            <Title title={'企業一覧'} />
+            <select className={clsx('w-100')}>
+              <option value='2021'>2021</option>
+              <option value='2022'>2022</option>
+            </select>
+          </div>
+          <div className={clsx('flex justify-end')}>
+            <RegistButton>
+              <RiAddCircleLine size={20} style={{ marginRight: 5 }} />
+              企業登録
+            </RegistButton>
+          </div>
+        </div>
+        <div className={clsx('mb-2 p-5')}>
+          <table className={clsx('mb-5 w-full table-fixed border-collapse')}>
+            <thead>
+              <tr
+                className={clsx('border border-x-white-0 border-b-primary-1 border-t-white-0 py-3')}
+              >
+                <th className={clsx('w-1/8 border-b-primary-1 pb-2')}>
+                  <div className={clsx('mr-1 text-center text-sm text-black-600')}>企業名</div>
+                </th>
+                <th className={clsx('w-1/8 border-b-primary-1 pb-2')}>
+                  <div className={clsx('text-center text-sm text-black-600')}>電話番号</div>
+                </th>
+                <th className={clsx('w-1/6 border-b-primary-1 pb-2')}>
+                  <div className={clsx('text-center text-sm text-black-600')}>メール</div>
+                </th>
+                <th className={clsx('w-1/6 border-b-primary-1 pb-2')}>
+                  <div className={clsx('text-center')}>住所</div>
+                </th>
+                <th className={clsx('w-1/8 border-b-primary-1 pb-2')}>
+                  <div className={clsx('text-center text-black-600')}>代表者</div>
+                </th>
+                <th className={clsx('w-1/12 border-b-primary-1 pb-2')}>
+                  <div className={clsx('text-center text-black-600')}></div>
+                </th>
+              </tr>
+            </thead>
+            <tbody className={clsx('border border-x-white-0 border-b-primary-1 border-t-white-0')}>
+              {sponsershipList.map((sponsorshipItem, index) => (
+                <tr key={sponsorshipItem.id}>
+                  <td
+                    className={clsx(
+                      'px-1',
+                      index === 0 ? 'pt-4 pb-3' : 'py-3',
+                      index === sponsershipList.length - 1 ? 'pb-4 pt-3' : 'border-b py-3',
+                    )}
+                  >
+                    <div className={clsx('text-center text-black-300')}>{sponsorshipItem.name}</div>
+                  </td>
+                  <td
+                    className={clsx(
+                      'px-1',
+                      index === 0 ? 'pt-4 pb-3' : 'py-3',
+                      index === sponsershipList.length - 1 ? 'pb-4 pt-3' : 'border-b py-3',
+                    )}
+                  >
+                    <div className={clsx('text-center text-black-300')}>{sponsorshipItem.tel}</div>
+                  </td>
+                  <td
+                    className={clsx(
+                      'px-1',
+                      index === 0 ? 'pt-4 pb-3' : 'py-3',
+                      index === sponsershipList.length - 1 ? 'pb-4 pt-3' : 'border-b py-3',
+                    )}
+                  >
+                    <div className={clsx('text-center text-black-300')}>
                       {sponsorshipItem.email}
-                    </Td>
-                    <Td>
-                      <Center color='black.300'>{sponsorshipItem.address}</Center>
-                    </Td>{' '}
-                    <Td>
-                      <Center color='black.300'>{sponsorshipItem.representative}</Center>
-                    </Td>
-                    <Td>
-                      <Center>
-                        <EditButton />
-                      </Center>
-                    </Td>
-                    <Td>
-                      <Center color='black.300'>{sponsorshipItem.created_at}</Center>
-                    </Td>
-                    <Td>
-                      <Center color='black.300'>{sponsorshipItem.updated_at}</Center>
-                    </Td>
-                  </Tr>
-                ))}
-              </Tbody>
-            </Table>
-          </Box>
-        </Box>
-      </Center>
-    </ChakraProvider>
+                    </div>
+                  </td>
+                  <td
+                    className={clsx(
+                      'px-1',
+                      index === 0 ? 'pt-4 pb-3' : 'py-3',
+                      index === sponsershipList.length - 1 ? 'pb-4 pt-3' : 'border-b py-3',
+                    )}
+                  >
+                    <div className={clsx('text-center text-black-300')}>
+                      {sponsorshipItem.address}
+                    </div>
+                  </td>{' '}
+                  <td
+                    className={clsx(
+                      'px-1',
+                      index === 0 ? 'pt-4 pb-3' : 'py-3',
+                      index === sponsershipList.length - 1 ? 'pb-4 pt-3' : 'border-b py-3',
+                    )}
+                  >
+                    <div className={clsx('text-center text-black-300')}>
+                      {sponsorshipItem.representative}
+                    </div>
+                  </td>
+                  <td
+                    className={clsx(
+                      'px-1',
+                      index === 0 ? 'pt-4 pb-3' : 'py-3',
+                      index === sponsershipList.length - 1 ? 'pb-4 pt-3' : 'border-b py-3',
+                    )}
+                  >
+                    <div className={clsx('text-center')}>
+                      <EditButton />
+                    </div>
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        </div>
+      </Card>
+    </MainLayout>
   );
 };
